ref(*): move data and handlers packages into pkg
