+++ filter_by_age_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/arschles/assert"
-	"github.com/deis/workflow-manager-api/data"
+	"github.com/deis/workflow-manager-api/pkg/data"
 	"github.com/deis/workflow-manager-api/rest"
 	"github.com/pborman/uuid"
 )
