+++ pkg/handlers/get_latest_versions.go
@@ -6,7 +6,7 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/deis/workflow-manager-api/data"
+	"github.com/deis/workflow-manager-api/pkg/data"
 	"github.com/deis/workflow-manager/types"
 	"github.com/jinzhu/gorm"
 )
