+++ pkg/handlers/handlers.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/deis/workflow-manager-api/data"
+	"github.com/deis/workflow-manager-api/pkg/data"
 	"github.com/deis/workflow-manager/types"
 	"github.com/gorilla/mux"
 	"github.com/jinzhu/gorm"
