+++ server_test.go
@@ -11,7 +11,7 @@ import (
 	"time"
 
 	"github.com/arschles/assert"
-	"github.com/deis/workflow-manager-api/data"
+	"github.com/deis/workflow-manager-api/pkg/data"
 	"github.com/deis/workflow-manager/types"
 	"github.com/jinzhu/gorm"
 )
